Add files via upload
Test module containing user defined functions that allow for a quit function and to both create a new bookshelf as well as append it with additional user submissions. The code can either be imported as a module or pasted into the user defined function section of the final script. Feel free to test and edit as needed!