Add files via upload
Version of book_data module with attempts at updated find_book() and export_to_csv() functions. find_books() was updated to utilize code from the update function, as both require similar first steps, and export was updated to open the json then open the csv file, so it should work by referencing the existing json file before trying to export. Make sure to change the name to just "book_data" when downloading this version so the import to the final script functions, date here is included so this file doesn't overwrite the original here